Backend/Spring/[kyh-p1] Spring Introduction/hello-spring/src/main/java/xyz/mon0mon/hellospring/repository/MemoryMemberRepository.java
package xyz.mon0mon.hellospring.repository;

import org.springframework.stereotype.Repository;
import xyz.mon0mon.hellospring.domain.Member;

import java.util.*;

public class MemoryMemberRepository implements MemberRepository {
    private static Map<Long, Member> store = new HashMap<>();
    private static Long sequence = 0L;

    @Override
    public Member save(Member member) {
        member.setId(++sequence);
        store.put(sequence, member);
        return member;
    }

    @Override
    public Optional<Member> findById(Long id) {
        return Optional.ofNullable(store.get(id));
    }

    @Override
    public Optional<Member> findByName(String name) {
        return store.values()
                    .stream()
                    .filter(member -> member.getName()
                                            .equals(name))
                    .findFirst();
    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(store.values());
    }

    public void clearStore() {
        store.clear();
    }
}
